+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+  - Include `limit` and `offset` in `thoth-client`'s works query
+  - Paginate `get_works` requests in export API using concurrent requests
+  - Input number of actix workers via CLI arguments
+
+### Added
+  - Work count query to `thoth-client`
 
 ## [[0.9.9]](https://github.com/thoth-pub/thoth/releases/tag/v0.9.9) - 2023-02-16
 ### Changed